+++ src/com/oltpbenchmark/benchmarks/wikipedia/procedures/AddWatchList.java
@@ -49,14 +49,14 @@ public class AddWatchList extends Procedure {
     // RUN
     // -----------------------------------------------------------------
 	
-    public void run(Connection conn, int userId, int nameSpace, String pageTitle) throws SQLException {
+    public void run(Connection conn, long userId, int nameSpace, String pageTitle) throws SQLException {
 		if (userId > 0) {
 		    // TODO: find a way to by pass Unique constraints in SQL server (Replace, Merge ..?)
 		    // Here I am simply catching the right excpetion and move on.
 		    try
 		    {
     			PreparedStatement ps = this.getPreparedStatement(conn, insertWatchList);
-    			ps.setInt(1, userId);
+    			ps.setLong(1, userId);
     			ps.setInt(2, nameSpace);
     			ps.setString(3, pageTitle);
     			ps.executeUpdate();
@@ -73,7 +73,7 @@ public class AddWatchList extends Procedure {
     				// if regular page, also add a line of
     				// watchlist for the corresponding talk page
     			    PreparedStatement ps = this.getPreparedStatement(conn, insertWatchList);
-    				ps.setInt(1, userId);
+    				ps.setLong(1, userId);
     				ps.setInt(2, 1);
     				ps.setString(3, pageTitle);
     				ps.executeUpdate();
@@ -86,7 +86,7 @@ public class AddWatchList extends Procedure {
 
 			PreparedStatement ps = this.getPreparedStatement(conn, setUserTouched);
 			ps.setString(1, TimeUtil.getCurrentTimeString14());
-			ps.setInt(2, userId);
+			ps.setLong(2, userId);
 			ps.executeUpdate();
 		}
 	}
